# docs/swagger.json
{
  "info": {
    "title": "1kbIdeas",
    "version": "1.0.0",
    "description": "The one-stop shop for all the self-help ideas you need",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
  "swagger": "2.0",
  "paths": {
...
    "/profile/:id": {
      "patch": {
        "description": "update user profile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userName",
            "description": "Username to update.",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstName",
            "description": "first name to update.",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastName",
            "description": "last name to update.",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "description": "email to update.",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "avaterUrl",
            "description": "image to update.",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "updated successfully."
          },
          "400": {
            "description": "not updated."
          }
        }
      },
      "get": {
        "description": "get a user profile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Username id.",
            "in": "body",
            "required": true,
            "type": "UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "request was successful."
          },
          "404": {
            "description": "user does not exist."
          }
        }
      }
    },
    "/profile/:userId/following": {
      "post": {
        "description": "follower a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Users id.",
            "in": "body",
            "required": true,
            "type": "UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully followed a user"
          },
          "404": {
            "description": "User does not exist."
          },
          "409": {
            "description": "User cannot perform this action."
          }
        }
      }
    },
    "/profile/:userId/followers": {
      "get": {
        "description": "get allfollowers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Username id.",
            "in": "body",
            "required": true,
            "type": "UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful."
          }
        }
      }
    },
...
    "/users/subscriptions": {
      "post": {
        "description": "subscribe to categories",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "application/json"
          },
          {
            "name": "categories",
            "description": "A stringified version of an array of categories",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/subscription"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "definitions": {
          "subscription": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
...
  "securityDefinitions": {},
...
}